+++ modules/i18n/src/models/namespaces/index.ts
@@ -71,3 +71,4 @@ export * from "./custom-authenticator-ns";
 export * from "./push-providers-ns";
 export * from "./webhooks-ns";
 export * from "./approval-workflows-ns";
+export * from "./agents-ns";